+++ db/dict
@@ -1,0 +1,98 @@
+Cisco
+Citrix
+AppFolio
+Seagate
+WD
+Fender
+Epiphone
+Gibson
+Amazon
+Tablet
+Lenovo
+Chromebook
+Toshiba
+HP
+KVM switch
+optical drive
+network card
+TV
+LCD
+raspberry pi
+steam
+ubisoft
+ea
+Midori
+Omas
+Edison
+Cross
+Delta
+Apica
+Waterman
+INTEL
+PHILIPS
+GRADO
+Kenwood
+Lite-on
+Asrock
+Gigabyte
+MSI
+Parker
+TWSBI
+Sheaffer
+Pilot
+Platinum
+Conklin
+Pelikan
+Diamine
+Rhodia
+Kaweco
+LAMY
+Stipula
+Clairefontaine
+motherboard
+cpu
+memory
+video card
+sound card
+keyboard
+monitor
+storage
+hard drive
+ssd
+apple
+microsoft
+acer
+asus
+sony
+HTC
+Samsung
+macbook
+surface pro
+kindle
+motorola
+dell
+smart phone
+laptop
+PBX
+network adapter
+router
+switch
+UPS
+sensor
+powercolor
+EVGA
+windows
+desktop
+linux
+unix
+coolermaster
+Nokia
+LG
+XBOX
+Play Station
+Nintendo
+Wii
+Bluetooth
+Logitech
+NVIDIA
+AMD
